LANGCHAIN_MODELS/requirements.txt
# Langchain-Core 

langchain
langchain-core

# OpenAI-Integration
langchain-openai 
openai
# Anthropic Integration
langchain-anthropic  

# Google Gemini (PaLM) Integration

langchain-google-genai
google-generativeai

# Hugging Face Integration 

langchain-huggingface
transformers>=4.35
huggingface-hub

# Environment variable Management 

python-dotenv

## Machine Learning Utilities 

numpy
scikit-learn